requirements.txt
aiogram
python-dotenv
python-binance
